Fix FindBugs warnings about StashBuildTrigger#job being null
Move startJob() and its dependencies from StashBuildTrigger to
StashRepository. StashBuildTrigger objects can have a null job field, but
it's impossible for StashRepository, which is instantiated once a
non-null job is passed to StashBuildTrigger#start().

Move corresponding unit tests from StashBuildTriggerTest to
StashRepositoryTest.
